delivery-image-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
pytesseract==0.3.13
